alpaca_stocks/requirements.txt
fastapi
uvicorn[standard]
alpaca-py
influxdb-client==1.40.0
requests==2.31.0
python-dotenv==1.0.1
